power: mmi-smbcharger-iio: Fix -Wpointer-to-int-cast.
Fixes:
drivers/power/mmi-smbcharger-iio/mmi-smbcharger-iio.c:4898:16: error:
    cast to smaller integer type 'u8' (aka 'unsigned char') from
    'const void *' [-Werror,-Wvoid-pointer-to-int-cast]
        int subtype = (u8)of_device_get_match_data(chip->dev);
                      ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

1 error generated.

Original-Change-Id: I30d500db49c4aae95efe7d41c88b04da3468fa29

+++ drivers/power/mmi-smbcharger-iio/mmi-smbcharger-iio.c
@@ -4895,7 +4895,7 @@ static int parse_mmi_dual_dt(struct smb_mmi_charger *chg)
 
 static int smb_mmi_chg_config_init(struct smb_mmi_charger *chip)
 {
-	int subtype = (u8)of_device_get_match_data(chip->dev);
+	int subtype = (u8)(long)of_device_get_match_data(chip->dev);
 
 	switch (subtype) {
 	case PM8150B:
